% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startStockfish.R
\name{startStockfish}
\alias{startStockfish}
\title{Start stockfish}
\usage{
startStockfish(...)
}
\description{
Starts a stockfish process from binaries included in the package. Functions
in stockfisher accept a stockfish process to keep things quick. If a process
is not provided they will spawn one of their own and close it after they are done.
This will add some overhead that.
}
\examples{
stockfish = startStockfish()
stockStep(posString = 'startpos',movetime = 1000,stockfish = stockfish)
stopStockfish()

}
