% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockStep.R
\name{stockStep}
\alias{stockStep}
\title{Title}
\usage{
stockStep(board, movetime = 5000, wtime = NULL, btime = NULL,
  depth = NULL, translate = TRUE, stockfish = NULL)
}
\arguments{
\item{board}{}

\item{movetime}{}

\item{translate}{}

\item{stockfish}{}
}
\description{
Title
}
