% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockStep.R
\name{stockStep}
\alias{stockStep}
\title{Calculate next move based on current position}
\usage{
stockStep(board = NULL, posString = NULL, movetime = 5000,
  wtime = NULL, btime = NULL, depth = NULL, translate = FALSE,
  ponder = FALSE, stockfish = NULL)
}
\arguments{
\item{board}{is an board generated by \code{rchess}. Mandatory if \code{translate} = TRUE}

\item{posString}{is a string that can be accepted by a UCI as an alternative to \code{board}
Will be ignored if \code{board} is defined}

\item{movetime}{How much time should the engine spend on thinking}

\item{wtime}{Alternative to \code{movetime}, must be provided with btime. How much time 
the white player has on the clock}

\item{btime}{Alternative to \code{movetime}, must be provided with wtime. How much time the black
player has on the clock.}

\item{depth}{Alternative to \code{movetime}. The search depth.}

\item{translate}{Should the move from the UCI be translated into standard chess notation. This format
is accepted by \code{rchess} boards. You need to provide a \code{board} for this option to work instead
of a \code{posString}}

\item{ponder}{Not implemented yet. Should the engine continue to ponder}

\item{stockfish}{A running UCI engine process. If NULL, a stockfish process will be spawned and stopped
at the end of function execution.}
}
\value{
a list containing 4 elements. \code{bestmove} is the calculated best move by the engine.
\code{ponder} is the opponent's best move against the engine's best move. \code{score} is engine's
perception of how good it's position is. \code{scoreType} is the type of score. If its cp, upperbound or
lowerbound, the score estimate is in centipawns. If it is mate, the score is a guess of how many turns till
the opponents demise (or the engine's if its negative).
}
\description{
Given a board or a position string that can be accepted by UCI, calculate the
next best move, opponents best move (ponder) and which player has the upper hand
(score)
}
